requirements.txt
streamlit==1.29.0
nltk==3.8.1
pdfminer.six==20221105
pickle-mixin==1.0.2
requests==2.31.0
streamlit-lottie==0.0.5